Add video resource to Docker section in README.md
A new step has been added to the "Using Docker Containers" section in the README.md file. This step instructs users to watch a video titled "The Logging Everyone Should Be Using in .NET" available on YouTube. This addition aims to provide users with additional resources to improve their understanding and implementation of logging within their .NET projects.